general, newlyhomeless, seniors dash to material
