+++ RCHIWebDash/RCHIWebDashBoard/src/components/charts/PieChart2.js
@@ -16,6 +16,7 @@ export default class PieChart2 extends React.Component {
             mydata : this.props.data,
             footer : this.props.footer,
             margin : this.props.margin,
+            position: this.props.position
         }
 
         //console.log("Pie chart")
@@ -28,13 +29,14 @@ export default class PieChart2 extends React.Component {
             mydata : this.props.data,
             footer : this.props.footer,
             margin : this.props.margin,
+            position: this.props.position
             
         })
     }
 
     render(){
         return (
-            <div style = {{height: this.props.height ? this.props.height : '100%', width: '100%', position:'absolute'}}>
+            <div style = {{height: this.props.height ? this.props.height : '100%', width: '100%', position: this.state.position ? this.state.position : 'absolute'}}>
             {/* // <div style={{height:this.state.height, width: '100%'}}> */}
             <ResponsivePie
                 data={this.props.data}
